add get request, and util functions + test code
Testing code accommodates for both nodejs and browser within the same
file. Applied the check whether it's browser-based or nodejs-based at
the top of the file in each testing code that must be ran on nodejs.

Insert the following code at the top of the file

var isBrowser = new Function("try {return this===window;}catch(e){
return false;}");
if (!isBrowser()) {
  bfet = require('../src/bfet');
}

Also loose glob file checking for testing source files in gulpfile and
jasmine (nodejs) to be
all the files in directory, not just root level.